Moved util functions from settings_service to app_utils
