+++ packages-exp/performance-exp/src/services/settings_service.ts
@@ -15,8 +15,6 @@
  * limitations under the License.
  */
 
-import { FirebaseApp } from '@firebase/app-types-exp';
-import { ERROR_FACTORY, ErrorCode } from '../utils/errors';
 import { mergeStrings } from '../utils/string_merger';
 
 let settingsServiceInstance: SettingsService | undefined;
@@ -66,28 +64,4 @@ export class SettingsService {
     }
     return settingsServiceInstance;
   }
-}
-
-export function getAppId(firebaseApp: FirebaseApp): string {
-  const appId = firebaseApp.options?.appId;
-  if (!appId) {
-    throw ERROR_FACTORY.create(ErrorCode.NO_APP_ID);
-  }
-  return appId;
-}
-
-export function getProjectId(firebaseApp: FirebaseApp): string {
-  const projectId = firebaseApp.options?.projectId;
-  if (!projectId) {
-    throw ERROR_FACTORY.create(ErrorCode.NO_PROJECT_ID);
-  }
-  return projectId;
-}
-
-export function getApiKey(firebaseApp: FirebaseApp): string {
-  const apiKey = firebaseApp.options?.apiKey;
-  if (!apiKey) {
-    throw ERROR_FACTORY.create(ErrorCode.NO_API_KEY);
-  }
-  return apiKey;
 }
